:boom: created system to send notifications to user groups with reading confirmation
Inside settings.py>TEMPLATES>OPTIONS>context_processors, a custom processor has been added to check if there are new notifications and if so, add the necessary JS (notify_user() function) to notify the user.
When the user will open or close the notification the js notification_read() function will be started which will make a HTTP Post request to the View ReadNotificationView.

The creation of a notification takes place by means of a special form that uses customized FilteredSelectMultiple field with the css and js of admin panel.

To represent a notification two models were created: Notification and RecipientsUser. The first model has an m2m relationship with UserModel through RecipientsUser. RecipientsUser will contain both recipient users of the notification and the reading state. Attention to the overriding of the form_valid() method inside the NotificationCreateView class, after saving the model you must call form.save_m2m() to enter the form's selected users inside RecipientsUser.

To send notifications via js it is necessary to set https to the site, for the local development of this project has been set https in the local django web server. Guide on how to do: https://web.dev/how-to-use-local-https/, https://github.com/FiloSottile/mkcert#installation, https://github.com/teddziuba/django-sslserver.
The web server can be started in two ways:
   1. Using the local IP of the network (I recommend making the ip static to avoid generating a different certificate each time)
      -> python3 manage.py runsslserver 0.0.0.0:8000 --certificate /path/to/ip.pem --key /path/to/ip-key.pem
   2. Using localhost
      -> python3 manage.py runsslserver --certificate /path/to/ip.pem --key /path/to/ip-key.pem