:boom: added all fields of the animal model
the Animal model has been completed and divided into 4 models, therefore 4 ModelForms have been created.
the models corresponding to CreateView and Updateview are united in a single model that does both.
the missing fields have been added to the template corresponding to the DetailView.
the new Animal model has 3 OneToOne fields. These fields cannot be shown with a simple ModelForm (because there are 4 models in total and not one only) and therefore they must have a special management.
the solution that has been implemented is the one that acts on the Views to control what is passed and received by the user by controlling the 4 forms simultaneously.

in the Animal model the delete() function has been overrided to apply an inverted CASCADE, ie if you cancel an animal then the connected models are also deleted.