:bug: denied cancellation of a group with users in
fixed #2
added check to prevent the cancellation of a group with users within it. To improve the site's usability a new View has been added that allows you to delete all users of a group in just two clicks.
corrected minor bugs in the same cancellation operation of area>boxes in the facility app